requirements.txt
langchain_core
langchain_community
langchain_openai
langchain
streamlit
python-dotenv
pandas
notebook
numpy
python-box
pyYAML
tqdm
ensure
types-PyYAML
scipy
reportlab
PyPDF2
markdown2
html2text  
faiss-cpu
duckduckgo-search
streamlit
pypdf
flask
uvicorn

-e .